+++ src/test/scala/com/chipprbots/ethereum/consensus/validators/std/StdSignedLegacyTransactionValidatorSpec.scala
@@ -237,7 +237,7 @@ class StdSignedLegacyTransactionValidatorSpec extends AnyFlatSpec with Matchers 
 
   it should "report as invalid a chain specific tx before eip155" taggedAs (UnitTest, ConsensusTest) in {
     val keyPair = crypto.generateKeyPair(new SecureRandom)
-    val stx = SignedTransaction.sign(txBeforeHomestead, keyPair, Some(0x03.toByte))
+    val stx = SignedTransaction.sign(txBeforeHomestead, keyPair, Some(BigInt(0x03)))
     StdSignedTransactionValidator.validate(
       stx,
       senderAccount = senderAccountAfterHomestead,
@@ -252,7 +252,7 @@ class StdSignedLegacyTransactionValidatorSpec extends AnyFlatSpec with Matchers 
 
   it should "report as valid a chain specific tx after eip155" taggedAs (UnitTest, ConsensusTest) in {
     val keyPair = crypto.generateKeyPair(new SecureRandom)
-    val stx = SignedTransaction.sign(txAfterHomestead, keyPair, Some(0x03.toByte))
+    val stx = SignedTransaction.sign(txAfterHomestead, keyPair, Some(BigInt(0x03)))
     StdSignedTransactionValidator.validate(
       stx,
       senderAccount = senderAccountAfterHomestead,
