+++ src/test/scala/com/chipprbots/ethereum/extvm/VMClientSpec.scala
@@ -222,7 +222,7 @@ class VMClientSpec extends AnyFlatSpec with Matchers with MockFactory {
       berlinBlockNumber = 0,
       mystiqueBlockNumber = 0,
       spiralBlockNumber = 0,
-      chainId = 0x3d.toByte
+      chainId = 0x3d
     )
     val evmConfig: EvmConfig = EvmConfig.FrontierConfigBuilder(blockchainConfigForEvm)
