+++ src/test/scala/com/chipprbots/ethereum/ledger/LedgerTestSetup.scala
@@ -254,7 +254,7 @@ trait DaoForkTestSetup extends TestSetup {
       )
     )
     .copy(
-      chainId = 0x01.toByte,
+      chainId = 0x01,
       networkId = 1,
       daoForkConfig = Some(supportDaoForkConfig),
       customGenesisFileOpt = None,
